# backend_svc/API.yaml
swagger: "2.0"
info:
  description: "This is a lab4-final project server."
  version: "1.0.0"
  title: "Listas Ya!"
host: "undefined"
basePath: "/"
...
schemes:
- "https"
paths:
...
  /lists/join/{userId}:
    post:
      tags: 
      - "lists"
      summary: "Add a user to a list"
      description: ""
      operationId: "joinList"
      produces: 
      - "application/json"
      parameters:
        - name: "userId"
          in: "path"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "successful operation"
        404:
          description: "user not found"
...